# Improve Int32/Int64 {Try}Parse performance for NumberStyles.Integer