Special-case the default NumberStyles.Integer format to improve the throughput of Int32/Int64.{Try}Parse.  Ports the implementation from Utf8Parser, modifying it to accomodate the additional cases needed for NumberStyles.Integer, e.g. culture impact on positive and negative signs, trailing nulls, leading and trailing whitespace, prohibiting additional characters at the end, etc.

Benchmarks:
```C#
using System;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Attributes.Jobs;
using BenchmarkDotNet.Running;

[MemoryDiagnoser]
[InProcess]
public class Benchmark
{
    private static void Main() => BenchmarkRunner.Run<Benchmark>();
    [Benchmark] public int Int32Parse1() => int.Parse("1");
    [Benchmark] public int Int32Parse12345() => int.Parse("12345");
    [Benchmark] public int Int32Parse1234567890() => int.Parse("1234567890");
    [Benchmark] public int Int32ParseN1() => int.Parse("-1");
    [Benchmark] public int Int32ParseW0W() => int.Parse(" 0 ");
    [Benchmark] public bool Int32TryParseInvalid() => int.TryParse("123a", out _);
    [Benchmark] public long Int64Parse1() => long.Parse("1");
    [Benchmark] public long Int64Parse123456789() => long.Parse("123456789");
    [Benchmark] public long Int64Parse1234567890123456789() => long.Parse("1234567890123456789");
    [Benchmark] public long Int64ParseN1() => long.Parse("-1");
    [Benchmark] public long Int64ParseW0W() => long.Parse(" 0 ");
    [Benchmark] public bool Int64TryParseInvalid() => long.TryParse("123a", out _);
}
```

Before/After:

| Benchmark                     | Before (ns) | After (ns) | Improvement | 
|-------------------------------|-------------|------------|-------------| 
|                   Int32Parse1 | 52.35       | 25.76      | 2.03x       | 
|               Int32Parse12345 | 64.41       | 28.43      | 2.27x       | 
|          Int32Parse1234567890 | 84.09       | 31.63      | 2.66x       | 
|                  Int32ParseN1 | 53.89       | 24.45      | 2.20x       | 
|                 Int32ParseW0W | 53.32       | 25.72      | 2.07x       | 
|          Int32TryParseInvalid | 61.60       | 34.84      | 1.77x       | 
|                   Int64Parse1 | 50.57       | 25.59      | 1.98x       | 
|           Int64Parse123456789 | 75.43       | 30.21      | 2.50x       | 
| Int64Parse1234567890123456789 | 110.70      | 37.86      | 2.92x       | 
|                  Int64ParseN1 | 51.81       | 24.47      | 2.12x       | 
|                 Int64ParseW0W | 51.77       | 25.65      | 2.02x       | 
|          Int64TryParseInvalid | 59.30       | 35.60      | 1.67x       | 

Contributes to https://github.com/dotnet/corefx/issues/30612
cc: @jkotas, @ahsonkhan, @danmosemsft