@danmosemsft 2018-07-13
Was test coverage already very good?