@stephentoub 2018-07-13
> Was test coverage already very good?

Yup.  But I also added some more tests in a corefx PR last night, in particular to improve the Int64 tests to the same level as the Int32 tests.