@stephentoub 2018-07-13
@dotnet-bot test OSX10.12 x64 Checked CoreFX Tests please
@dotnet-bot test Windows_NT x64 Checked CoreFX Tests please