@stephentoub 2018-07-16
(As part of porting/implementing the unsigned equivalents, I found a bug in this code, where if an input has both an overflow and a format error, we may throw the wrong exception, based on compat with netfx... I'll fix it as part of this next set of changes.)